Refactor

+++ PowershellTools.HostService/ServiceManagement/Debugging/PowershellDebuggingService.cs
@@ -140,32 +140,35 @@ namespace PowerShellTools.HostService.ServiceManagement.Debugging
             ServiceCommon.Log("Setting breakpoint ...");
             try
             {
-                using (var pipeline = (_runspace.CreatePipeline()))
+                if (_runspace.RunspaceAvailability == RunspaceAvailability.Available)
                 {
-                    var command = new Command("Set-PSBreakpoint");
-
-                    string file = bp.ScriptFullPath;
-                    if (_runspace.ConnectionInfo != null && _mapLocalToRemote.ContainsKey(bp.ScriptFullPath))
+                    using (var pipeline = (_runspace.CreatePipeline()))
                     {
-                        file = _mapLocalToRemote[bp.ScriptFullPath];
+                        var command = new Command("Set-PSBreakpoint");
+
+                        string file = bp.ScriptFullPath;
+                        if (_runspace.ConnectionInfo != null && _mapLocalToRemote.ContainsKey(bp.ScriptFullPath))
+                        {
+                            file = _mapLocalToRemote[bp.ScriptFullPath];
+                        }
+
+                        command.Parameters.Add("Script", file);
+
+                        command.Parameters.Add("Line", bp.Line);
+
+                        pipeline.Commands.Add(command);
+
+                        breakpoints = pipeline.Invoke();
                     }
 
-                    command.Parameters.Add("Script", file);
-
-                    command.Parameters.Add("Line", bp.Line);
-
-                    pipeline.Commands.Add(command);
-
-                    breakpoints = pipeline.Invoke();
-                }
-
-                var pobj = breakpoints.FirstOrDefault();
-                if (pobj != null)
-                {
-                    _psBreakpointTable.Add(
-                        new PowerShellBreakpointRecord(
-                            bp,
-                            ((LineBreakpoint)pobj.BaseObject).Id));
+                    var pobj = breakpoints.FirstOrDefault();
+                    if (pobj != null)
+                    {
+                        _psBreakpointTable.Add(
+                            new PowerShellBreakpointRecord(
+                                bp,
+                                ((LineBreakpoint)pobj.BaseObject).Id));
+                    }
                 }
             }
             catch (InvalidOperationException)
@@ -184,28 +187,31 @@ namespace PowerShellTools.HostService.ServiceManagement.Debugging
 
             if (id >= 0)
             {
-                ServiceCommon.Log("Removing breakpoint ...");
-
-                using (var pipeline = (_runspace.CreatePipeline()))
+                if (_runspace.RunspaceAvailability == RunspaceAvailability.Available)
                 {
-                    var command = new Command("Remove-PSBreakpoint");
+                    ServiceCommon.Log("Removing breakpoint ...");
 
-                    string file = bp.ScriptFullPath;
-                    if (_runspace.ConnectionInfo != null && _mapLocalToRemote.ContainsKey(bp.ScriptFullPath))
+                    using (var pipeline = (_runspace.CreatePipeline()))
                     {
-                        file = _mapLocalToRemote[bp.ScriptFullPath];
+                        var command = new Command("Remove-PSBreakpoint");
+
+                        string file = bp.ScriptFullPath;
+                        if (_runspace.ConnectionInfo != null && _mapLocalToRemote.ContainsKey(bp.ScriptFullPath))
+                        {
+                            file = _mapLocalToRemote[bp.ScriptFullPath];
+                        }
+
+                        command.Parameters.Add("Id", id);
+
+                        pipeline.Commands.Add(command);
+
+                        pipeline.Invoke();
                     }
 
-                    command.Parameters.Add("Id", id);
-
-                    pipeline.Commands.Add(command);
-
-                    pipeline.Invoke();
-                }
-
-                foreach (var p in _psBreakpointTable.Where(b => b.PSBreakpoint.Equals(bp)))
-                {
-                    _psBreakpointTable.Remove(p);
+                    foreach (var p in _psBreakpointTable.Where(b => b.PSBreakpoint.Equals(bp)))
+                    {
+                        _psBreakpointTable.Remove(p);
+                    }
                 }
             }
         }
@@ -220,25 +226,28 @@ namespace PowerShellTools.HostService.ServiceManagement.Debugging
 
             if (id >= 0)
             {
-                ServiceCommon.Log(string.Format("{0} breakpoint ...", enable ? "Enable" : "Disable"));
-
-                using (var pipeline = (_runspace.CreatePipeline()))
+                if (_runspace.RunspaceAvailability == RunspaceAvailability.Available)
                 {
-                    string cmd = enable ? "Enable-PSBreakpoint" : "Disable-PSBreakpoint";
+                    ServiceCommon.Log(string.Format("{0} breakpoint ...", enable ? "Enable" : "Disable"));
 
-                    var command = new Command(cmd);
-
-                    string file = bp.ScriptFullPath;
-                    if (_runspace.ConnectionInfo != null && _mapLocalToRemote.ContainsKey(bp.ScriptFullPath))
+                    using (var pipeline = (_runspace.CreatePipeline()))
                     {
-                        file = _mapLocalToRemote[bp.ScriptFullPath];
+                        string cmd = enable ? "Enable-PSBreakpoint" : "Disable-PSBreakpoint";
+
+                        var command = new Command(cmd);
+
+                        string file = bp.ScriptFullPath;
+                        if (_runspace.ConnectionInfo != null && _mapLocalToRemote.ContainsKey(bp.ScriptFullPath))
+                        {
+                            file = _mapLocalToRemote[bp.ScriptFullPath];
+                        }
+
+                        command.Parameters.Add("Id", id);
+
+                        pipeline.Commands.Add(command);
+
+                        pipeline.Invoke();
                     }
-
-                    command.Parameters.Add("Id", id);
-
-                    pipeline.Commands.Add(command);
-
-                    pipeline.Invoke();
                 }
             }
             else
@@ -254,24 +263,27 @@ namespace PowerShellTools.HostService.ServiceManagement.Debugging
         {
             ServiceCommon.Log("ClearBreakpoints");
 
-            IEnumerable<PSObject> breakpoints;
-
-            using (var pipeline = (_runspace.CreatePipeline()))
+            if (_runspace.RunspaceAvailability == RunspaceAvailability.Available)
             {
-                var command = new Command("Get-PSBreakpoint");
-                pipeline.Commands.Add(command);
-                breakpoints = pipeline.Invoke();
-            }
+                IEnumerable<PSObject> breakpoints;
 
-            if (!breakpoints.Any()) return;
+                using (var pipeline = (_runspace.CreatePipeline()))
+                {
+                    var command = new Command("Get-PSBreakpoint");
+                    pipeline.Commands.Add(command);
+                    breakpoints = pipeline.Invoke();
+                }
 
-            using (var pipeline = (_runspace.CreatePipeline()))
-            {
-                var command = new Command("Remove-PSBreakpoint");
-                command.Parameters.Add("Breakpoint", breakpoints);
-                pipeline.Commands.Add(command);
+                if (!breakpoints.Any()) return;
 
-                pipeline.Invoke();
+                using (var pipeline = (_runspace.CreatePipeline()))
+                {
+                    var command = new Command("Remove-PSBreakpoint");
+                    command.Parameters.Add("Breakpoint", breakpoints);
+                    pipeline.Commands.Add(command);
+
+                    pipeline.Invoke();
+                }
             }
         }
 
@@ -307,6 +319,8 @@ namespace PowerShellTools.HostService.ServiceManagement.Debugging
         public bool Execute(string commandLine)
         {
             ServiceCommon.Log("Start executing ps script ...");
+            
+            bool commandExecuted = false;
 
             try
             {
@@ -347,6 +361,8 @@ namespace PowerShellTools.HostService.ServiceManagement.Debugging
                 {
                     lock (ServiceCommon.RunspaceLock)
                     {
+                        commandExecuted = true;
+
                         using (_currentPowerShell = PowerShell.Create())
                         {
                             _currentPowerShell.Runspace = _runspace;
@@ -367,17 +383,22 @@ namespace PowerShellTools.HostService.ServiceManagement.Debugging
             {
                 ServiceCommon.Log("Type,  Exception: {0}", ex.Message);
                 OnTerminatingException(ex);
+
                 return false;
             }
             catch (Exception ex)
             {
                 ServiceCommon.Log("Terminating error,  Exception: {0}", ex.Message);
                 OnTerminatingException(ex);
+                
                 return false;
             }
             finally
             {
-                DebuggerFinished();
+                if (commandExecuted)
+                {
+                    DebuggerFinished();
+                }
             }
         }
